replace type activations with value activations
The existing value activations already has all the functionality needed,
so the dedicated activations for types are not needed.

The use of variables instead of just types allows recording positions
and access for type declarations.

Also, this simplifies the code, because no additional variables have to
be created when recording occurrences, and the import checkign for
values and types can now be generic for variables.