Update CCF encoder to use DictionaryType directly
Dictionary.DictionaryType used to be of type cadence.Type, which
requires type assertion during encoding.  Now Dictionary.DictionaryType
is of type *cadence.DictionaryType, and type assertion is no longer
needed.

Since all cadence.Value's runtime type is checked to be non-nil in
encodeValue(), v.DictionaryType is not nil in encodeDictionary().