Remove cadence-type-id from restricted-type(-value)
In JSON-CDC, sometimes cadence-type-id was encoded when it was just
a "stringification" of other encoded data and not necessary to encode.

In CCF, we don't need to keep this inefficiency for the sake of compatibility.

This commit removes cadence-type-id from restricted-type and
restricted-type-value.

It also modifies RestrictedType to generate ID if it is empty.  Generated ID is
stringification of its type and restrictions.