Add reproducer for false alarm (unreachable slab)
Cadence smoke test is reporting false alarm about
"slab was not reachable".

To summarize, storage.CheckHealth() should be called after
array.DeepRemove(), rather than in the middle of array.DeepRemove().

CheckHealth() is called in the middle of array.DeepRemove() when:
- array.DeepRemove() calls childArray1 and childArray2 DeepRemove()
- DeepRemove() calls maybeValidateAtreeValue()
- maybeValidateAtreeValue() calls CheckHealth()

The details are more complex than this oversimplified summary.

For more context, see comments/discussion on GitHub at
https://github.com/onflow/cadence/pull/2882#issuecomment-1781298107