Remove enum type checking when removing unreferenced slab
Currently enum type is the only composite type that needs to be
removed explicitly when used as key to update existing dictionary
element.

However, (as noted by Supun) Cadence might in the future add
support for other composite types as key.

This commit removes enum type checking to support new composite
key types to be added to Cadence language in the future.